requirements.txt
pandas==1.5.2
python-telegram-bot==20.3
sklearn==0.0.post1
typing_extensions==4.5.0